refactor(architecture): move baseline logic to engine and test extension
