+++ src/core/extension/test/parsing.rs
@@ -1,7 +1,7 @@
 use serde::Serialize;
 
 use crate::test_analyze::{TestAnalysis, TestAnalysisInput};
-use crate::test_baseline::TestCounts;
+use crate::extension::test::TestCounts;
 use crate::utils::io;
 use crate::utils::output_parse::{Aggregate, DeriveRule, ParseRule, ParseSpec};
