+++ src/core/mod.rs
@@ -33,7 +33,6 @@ pub mod ssh;
 pub mod scaffold;
 pub mod symbol_graph;
 pub mod test_analyze;
-pub mod test_baseline;
 pub mod test_drift;
 pub mod test_workflow;
 pub mod undo;
